reclone the repo and add 2 cpe solution

+++ CPE/1star/39. Fourth Point/39_Fourth_point.cpp
@@ -4,11 +4,11 @@ using namespace std;
 
 int main(void)
 {
-    float a1, a2, a3, a4, b1, b2, b3, b4;
+    double a1, a2, a3, a4, b1, b2, b3, b4;
     while(cin >> a1 >> b1 >> a2 >> b2 >> a3 >> b3 >> a4 >> b4)
     {   
         // find which 2 point are the same
-        float sameX, sameY, diff1x, diff1y, diff2x, diff2y;
+        double sameX, sameY, diff1x, diff1y, diff2x, diff2y;
         if (a1 == a3 && b1 == b3)
         {
             sameX = a1; sameY = b1;
